Fixing points
Remove half points, adding full points.

+++ src/Model.cs
@@ -100,7 +100,7 @@ namespace Ratcycle
 			/// <summary>
 			/// Types of garbage
 			/// </summary>
-            public static int points;
+            public static float points;
 			public enum Type { Normal, Strong }
 			public enum Category { Plastic, Paper, Chemical, Green, Other }
 		}
